Restoring grammar back to include human-readable operators
Initially I thought that I would introduce textual operators (gt, lt,
etc.), as most of the common logical and conditional operator symbols
are reserved characters according to RFC 3986. That would have meant
that an end user, entering the query as a REST GET query parameter
argument would have to percent encode these reserved characters. I
initially wanted to avoid having the user do that but I am coming around
to the idea that that's ok. The core audience for resql is likely not an
end-user, but other software applications. However, even for end-users
it's not that hard to encode a URI component at the click of a button,
like in the Postman interface.